requirements.txt
streamlit
Pillow
numpy
torch
pytorch_lightning
torchvision
tqdm
einops
ema_pytorch
